+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 dev
 ---
 
-* -
+* Added a new option: `escape_html`. It escapes the `<`, `>`, and `&` HTML
+  characters in notification messages.
 
 0.1 (2015-06-21)
 ----------------